+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Pvp stats summary
 - Progress bar to allow custom label
 - Pvp match component, adds progress bar, modifies styles
+- Favourite profession to user page
 
 ## [2.12.0] - 2016/10/29
 ### Added