+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Pvp stats summary
 - Progress bar to allow custom label
 - Pvp match component, adds progress bar, modifies styles
+
+### Added
 - Favourite profession to user page
 
 ## [2.12.0] - 2016/10/29